Add mail microservice

+++ backend/microservices/cars/src/api/carApi.js
@@ -2,45 +2,46 @@ const service = require('../service/carService');
 
 module.exports = function(app){
 
-    app.get('/api/cars/v1/get/:id', async (req, res) => {
+    app.get('/cars/get/:id', async (req, res) => {
         console.log(req.method + req.route.path);
+
+        if(!req.params.id) return res.status('400');
+
         let result = await service.get(req.params.id);
         res.status(result.status).send(result.response);
     });
 
-    app.post('/api/cars/v1/create', async (req, res) => {
+    app.post('/cars/create', async (req, res) => {
         console.log(req.method + req.route.path);
+
+        if(!req.body) return res.status('400');
+
         let result = await service.create(req.body);
         res.status(result.status).send(result.response);
     });
 
-    app.put('/api/cars/v1/update', async (req, res) => {
+    app.put('/cars/update', async (req, res) => {
         console.log(req.method + req.route.path);
+
+        if(!req.body) return res.status('400');
+
         let result = await service.update(req.body);
         res.status(result.status).send(result.response);
     });
 
-    app.delete('/api/cars/v1/remove/:id', async (req, res) => {
+    app.delete('/cars/remove/:id', async (req, res) => {
         console.log(req.method + req.route.path);
+
+        if(!req.params.id) return res.status('400');
+
         let result = await service.remove(req.params.id);
         res.status(result.status).send(result.response);
     });
 
-    app.get('/api/cars/v1/all', async (req, res) => {
+    app.get('/cars/all', async (req, res) => {
         console.log(req.method + req.route.path);
         let result = await service.getAll();
         res.status(result.status).send(result.response);
     });
 
-    app.post('/api/cars/v1/rented', async (req, res) => {
-        console.log(req.method + req.route.path);
-        let result = await service.rented(req.body);
-        res.status(result.status).send(result.response);
-    });
-    app.get('/api/cars/v1/stats', async (req, res) => {
-        console.log(req.method + req.route.path);
-        let result = await service.stats();
-        res.status(result.status).send(result.response);
-    });
 };
-
